mcp-client/shopify-sales-agent/pyproject.toml
[project]
name = "shopify-sales-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.1",
    "bs4>=0.0.2",
    "gradio>=5.33.0",
    "mcp>=1.9.2",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "strands-agents>=0.1.6",
    "strands-agents-tools>=0.1.4",
    "pytest>=8.4.0",
    "mcp-server-time>=0.6.2",
    "mcp-shopify-products",
]

[tool.uv.sources]
mcp-shopify-products = { path = "../../mcp-server/mcp-shopify-products" }
